align FLOAT mutation values with float64 within Row/Json converters
RowToMutationConverter declared FLOAT columns as float64 in
convertFieldType and wrote float64 arrays, but wrote scalar FLOAT values
as float32 - Spanner rejects the mutation against the schema the same
converter generates. JsonToMutationConverter wrote scalar FLOAT as
float64 but arrays as float32Array. Both now write float64 consistently.
(AvroToMutationConverter is consistently float32 end-to-end and is left
unchanged.)

Verified fail-first; regression assertions updated.

+++ src/test/java/com/mercari/solution/util/schema/converter/RowToMutationConverterTest.java
@@ -109,7 +109,8 @@ public class RowToMutationConverterTest {
         Assertions.assertEquals(Value.int64(16L), map.get("int16Field"));
         Assertions.assertEquals(Value.int64(32L), map.get("int32Field"));
         Assertions.assertEquals(Value.int64(64L), map.get("int64Field"));
-        Assertions.assertEquals(Value.float32(1.5f), map.get("floatField"));
+        // FLOAT is written as float64, matching convertFieldType (Type.float64) and the array path
+        Assertions.assertEquals(Value.float64(1.5d), map.get("floatField"));
         Assertions.assertEquals(Value.float64(2.5d), map.get("doubleField"));
         Assertions.assertEquals(Value.numeric(new BigDecimal("123.45")), map.get("decimalField"));
         Assertions.assertEquals(Value.timestamp(Timestamp.parseTimestamp("2024-03-01T12:00:00Z")), map.get("datetimeField"));
